+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 
 ## [Unreleased]
 
-## [0.8.0] - 2026-09-03
+## [0.8.0] - 2026-09-04
 
 ### Added